Fix access to base application page.
This is a bug or limitation in Horde_Routes, that makes the whole
sub-path of a URL optional if you specify a default value for one of the
path parameters. E.g. we specify a default of "docs" for the :file
parameter here, which is turned into the regexp "(.../(docs/...)?)?".
This makes the complete end of the path optional, i.e. "/docs" is no
longer required. That's why this rule matches any "/apps/:app" URL.